requirements.txt
discord.py~=2.5.2
PyYAML~=6.0.2
aiohttp~=3.12.9
requests~=2.32.3
steam~=1.4.4